partial version miss one case : intersection on different points

+++ src/jdis/jdis.h
@@ -1,3 +1,3 @@
 #include "../avl/bst.h"
 
-float bst_jdis(char *p1, char *p2);
+double bst_jdis(char *p1, char *p2);
